+++ Cargo.toml
@@ -187,7 +187,8 @@ nearly-linear = { path = "crates/utils/nearly-linear", version = "0.3" }
 # Dependencies maintained by Sovereign
 jmt = { version = "0.12.0", default-features = false }
 reltester = { git = "https://github.com/sovereign-labs/reltester.git", rev = "d6209abff72929f5ba41614f5cea3e4a748428c3" }
-rockbound = { git = "https://github.com/sovereign-labs/rockbound", rev = "bb21a29cee84d4491850acd553b150ec6cc1ca71" }
+# rockbound = { git = "https://github.com/sovereign-labs/rockbound", rev = "bb21a29cee84d4491850acd553b150ec6cc1ca71" }
+rockbound = { path = "../rockbound"}
 nmt-rs = { version = "0.2.5", features = ["serde", "borsh"] }
 # External dependencies
 nomt-core = { version = "1.0.2", default-features = false, features = ["borsh", "serde"] }
