+++ crates/full-node/sov-db/src/accessory_db.rs
@@ -26,7 +26,6 @@ impl AccessoryDb {
             name: Self::DB_NAME,
             path_suffix: Self::DB_PATH_SUFFIX,
             columns: ACCESSORY_TABLES.to_vec(),
-            cacheable_columns: vec![],
         }
     }
 
@@ -38,7 +37,7 @@ impl AccessoryDb {
     /// Queries for a value in the [`AccessoryDb`], given a key.
     pub fn get_value_option(
         &self,
-        key: &AccessoryKey,
+        key: &SlotKey,
         version: SlotNumber,
     ) -> anyhow::Result<AccessoryStateValue> {
         ensure_version_is_correct(
