+++ crates/full-node/sov-db/src/flat_db.rs
@@ -3,13 +3,12 @@
 
 use std::sync::Arc;
 
-use rockbound::Schema;
 use rockbound::{
     default_cf_descriptor, rocksdb::ColumnFamilyDescriptor, versioned_db::VersionedDB, SchemaBatch,
 };
 
+use rockbound::versioned_db::SchemaWithVersion;
 use crate::metrics::nomt::FlatStateCommitMetric;
-use crate::schema::namespace::{NomtCommittedVersion, NomtHistoricalState, NomtPruningState};
 use crate::{
     historical_state::StateChanges,
     namespaces::{KernelNamespace, UserNamespace},
@@ -19,19 +18,14 @@ use crate::{
 
 /// A database to store the flat state of the rollup (i.e. the raw key-value pairs)
 pub struct FlatStateDb {
-    pub(crate) user: VersionedDB<NomtStateValues<UserNamespace>>,
-    pub(crate) kernel: VersionedDB<NomtStateValues<KernelNamespace>>,
+    pub(crate) user: Arc<VersionedDB<NomtStateValues<UserNamespace>>>,
+    pub(crate) kernel: Arc<VersionedDB<NomtStateValues<KernelNamespace>>>,
     pub(crate) other: Arc<rockbound::DB>,
     #[allow(dead_code)]
     // We don't technically need to store the archival db here - it's only accessed through the user/kernel versioned DB wrappers.
     // We keep it here so that the internal structure is more legible by glancing at this struct. Note that unless the user has set the config to separate out the archival db,
     // this will point to the same rockbound::DB as the `other` field.
     pub(crate) archival: Arc<rockbound::DB>,
-}
-
-pub struct FlatDbCommitData {
-    pub archival: SchemaBatch,
-    pub flat: SchemaBatch,
 }
 
 impl FlatStateDb {
@@ -56,31 +50,33 @@ impl FlatStateDb {
             separate_archival,
         )?;
         let other = Self::get_rockbound_options(columns)
-            .setup_db_in_path_with_column_descriptors(path.clone(), cache_size)?;
+            .setup_db_in_path_with_column_descriptors(path.clone())?;
         let other = Arc::new(other);
         let archival = if separate_archival {
             let archival_path = path.join(Self::ARCHIVAL_DB_PATH_SUFFIX);
             let archival_columns = vec![
-                default_cf_descriptor(NomtHistoricalState::<UserNamespace>::COLUMN_FAMILY_NAME),
-                default_cf_descriptor(NomtHistoricalState::<KernelNamespace>::COLUMN_FAMILY_NAME),
-                default_cf_descriptor(NomtPruningState::<UserNamespace>::COLUMN_FAMILY_NAME),
-                default_cf_descriptor(NomtPruningState::<KernelNamespace>::COLUMN_FAMILY_NAME),
-                default_cf_descriptor(NomtCommittedVersion::<UserNamespace>::COLUMN_FAMILY_NAME),
-                default_cf_descriptor(NomtCommittedVersion::<KernelNamespace>::COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<UserNamespace>::HISTORICAL_COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<KernelNamespace>::HISTORICAL_COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<UserNamespace>::PRUNING_COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<KernelNamespace>::PRUNING_COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<UserNamespace>::VERSION_METADATA_COLUMN_FAMILY_NAME),
+                default_cf_descriptor(NomtStateValues::<KernelNamespace>::VERSION_METADATA_COLUMN_FAMILY_NAME),
             ];
             let archival = Self::get_rockbound_options(archival_columns);
-            Arc::new(archival.setup_db_in_path_with_column_descriptors(archival_path, 0)?)
+            Arc::new(archival.setup_db_in_path_with_column_descriptors(archival_path)?)
         } else {
             other.clone()
         };
-        let user = VersionedDB::<NomtStateValues<UserNamespace>>::from_dbs(
+        let user = Arc::new(VersionedDB::<NomtStateValues<UserNamespace>>::from_dbs(
             other.clone(),
             archival.clone(),
-        )?;
-        let kernel = VersionedDB::<NomtStateValues<KernelNamespace>>::from_dbs(
+            cache_size,
+        )?);
+        let kernel = Arc::new(VersionedDB::<NomtStateValues<KernelNamespace>>::from_dbs(
             other.clone(),
             archival.clone(),
-        )?;
+            100_000
+        )?);
         Ok(Self {
             user,
             kernel,
@@ -95,12 +91,12 @@ impl FlatStateDb {
     }
 
     /// Get the underlying [`VersionedDB`] for the user state.
-    pub fn get_user_db(&self) -> &VersionedDB<NomtStateValues<UserNamespace>> {
+    pub fn get_user_db(&self) -> &Arc<VersionedDB<NomtStateValues<UserNamespace>>> {
         &self.user
     }
 
     /// Get the underlying [`VersionedDB`] for the kernel state.
-    pub fn get_kernel_db(&self) -> &VersionedDB<NomtStateValues<KernelNamespace>> {
+    pub fn get_kernel_db(&self) -> &Arc<VersionedDB<NomtStateValues<KernelNamespace>>> {
         &self.kernel
     }
 
@@ -108,31 +104,55 @@ impl FlatStateDb {
     pub fn get_rockbound_options(
         columns: Vec<ColumnFamilyDescriptor>,
     ) -> DbOptions<ColumnFamilyDescriptor> {
-        let cacheable_columns = vec![
-            NomtCommittedVersion::<UserNamespace>::COLUMN_FAMILY_NAME.to_string(),
-            NomtCommittedVersion::<KernelNamespace>::COLUMN_FAMILY_NAME.to_string(),
-            NomtStateValues::<UserNamespace>::COLUMN_FAMILY_NAME.to_string(),
-            NomtStateValues::<KernelNamespace>::COLUMN_FAMILY_NAME.to_string(),
-        ];
         DbOptions {
             name: Self::DB_NAME,
             path_suffix: Self::DB_PATH_SUFFIX,
             columns,
-            cacheable_columns,
         }
     }
 
-    /// Coalesce all the changes into a single schema batch.
-    /// Assumption: only a single thread is committing at a time. Calling prepare_commit multiple times
-    /// will result in a version mismatch.
-    fn prepare_commit(&self, state: StateChanges) -> anyhow::Result<FlatDbCommitData> {
-        let StateChanges {
-            user,
-            kernel,
-            other,
-        } = state;
+    // /// Coalesce all the changes into a single schema batch.
+    // /// Assumption: only a single thread is committing at a time. Calling prepare_commit multiple times
+    // /// will result in a version mismatch.
+    // fn prepare_commit(&self, state: StateChanges) -> anyhow::Result<FlatDbCommitData> {
+    //     let StateChanges {
+    //         user,
+    //         kernel,
+    //         other,
+    //     } = state;
 
-        let mut other_changes = Arc::try_unwrap(other).unwrap_or_else(|arc| (*arc).clone());
+    //     let mut other_changes = Arc::try_unwrap(other).unwrap_or_else(|arc| (*arc).clone());
+    //     let version = self
+    //         .kernel
+    //         .get_committed_version()?
+    //         .and_then(|v| v.checked_add(1))
+    //         .unwrap_or(0);
+    //     if cfg!(debug_assertions) {
+    //         let user_version = self
+    //             .user
+    //             .get_committed_version()?
+    //             .and_then(|v| v.checked_add(1))
+    //             .unwrap_or(0);
+    //         assert_eq!(user_version, version);
+    //     }
+    //     let mut archival_data = SchemaBatch::default();
+    //     self.user
+    //         .materialize(&user, &mut other_changes, &mut archival_data, version)?;
+    //     self.kernel
+    //         .materialize(&kernel, &mut other_changes, &mut archival_data, version)?;
+
+    //     Ok(FlatDbCommitData {
+    //         archival: archival_data,
+    //         flat: other_changes,
+    //     })
+    // }
+
+    /// Coalesce all the changes into a single schema batch and write it atomically.
+    pub fn commit(&self, state: StateChanges) -> anyhow::Result<FlatStateCommitMetric> {
+        let start_prepare = std::time::Instant::now();
+        // let commit = self.prepare_commit(state)?;
+        let prepare = start_prepare.elapsed();
+        let start_write = std::time::Instant::now();
         let version = self
             .kernel
             .get_committed_version()?
@@ -146,30 +166,14 @@ impl FlatStateDb {
                 .unwrap_or(0);
             assert_eq!(user_version, version);
         }
-        let mut archival_data = SchemaBatch::default();
-        self.user
-            .materialize(&user, &mut other_changes, &mut archival_data, version)?;
-        self.kernel
-            .materialize(&kernel, &mut other_changes, &mut archival_data, version)?;
-
-        Ok(FlatDbCommitData {
-            archival: archival_data,
-            flat: other_changes,
-        })
-    }
-
-    /// Coalesce all the changes into a single schema batch and write it atomically.
-    pub fn commit(&self, state: StateChanges) -> anyhow::Result<FlatStateCommitMetric> {
-        let start_prepare = std::time::Instant::now();
-        let commit = self.prepare_commit(state)?;
-        let prepare = start_prepare.elapsed();
-        let start_write = std::time::Instant::now();
-        self.archival.write_schemas(commit.archival)?;
+        self.kernel.commit(&state.kernel, version)?;
+        self.user.commit(&state.user, version)?;
+        // self.archival.write_schemas(commit.archival)?;
         #[cfg(feature = "test-utils")]
         if cfg!(debug_assertions) && std::env::var("SOV_CRASH_ON_COMMIT").is_ok() {
             panic!("SOV_CRASH_ON_COMMIT is set, crashing the node");
         }
-        self.other.write_schemas(commit.flat)?;
+        self.other.write_schemas(&state.other)?;
         let write = start_write.elapsed();
         Ok(FlatStateCommitMetric { prepare, write })
     }
