+++ crates/full-node/sov-db/src/schema/types.rs
@@ -10,6 +10,9 @@ use sov_rollup_interface::stf::{
 };
 use sov_rollup_interface::Bytes;
 
+/// Defines slot-key and slot-value types, which can be cheaply cloned with minimal allocation.
+pub mod slot_key;
+
 /// A cheaply cloneable bytes abstraction for use within the trust boundary of the node
 /// (i.e. when interfacing with the database). Serializes and deserializes more efficiently,
 /// than most bytes abstractions, but is vulnerable to out-of-memory attacks
